+++ r5dev/vgui/vgui_debugpanel.cpp
@@ -94,18 +94,19 @@ void CTextOverlay::DrawNotify(void)
 		if (flTimeleft < 1.0f)
 		{
 			float f = clamp(flTimeleft, 0.0f, 1.0f) / 1.0f;
-			c[3] = static_cast<int>(f * 255.0f);
+			c[3] = int(f * 255.0f);
 
 			if (i == 0 && f < 0.2f)
 			{
-				y -= m_nFontHeight * (static_cast<float>(1.0f - f / 0.2f));
+				y -= int(m_nFontHeight * (float(1.0f - f / 0.2f)));
 			}
 		}
 		else
 		{
 			c[3] = 255;
 		}
-		CMatSystemSurface_DrawColoredText(g_pMatSystemSurface, v_Rui_GetFontFace(), m_nFontHeight, x, y, c.r(), c.g(), c.b(), c.a(), m_vNotifyText[i].m_svMessage.c_str());
+		CMatSystemSurface_DrawColoredText(g_pMatSystemSurface, v_Rui_GetFontFace(),
+			m_nFontHeight, x, y, c.r(), c.g(), c.b(), c.a(), m_vNotifyText[i].m_svMessage.c_str());
 
 		if (IsX360())
 		{
